Added GoogleSearch class.
It seems that the google-search gem uses the official Google Search API.
But this gem does not fit our needs, because the API limits the number of
results to 64, and for us purposes we need at least 200 responses.

Because of this, this gem is not a dependency anymore. Instead, I'll going
to write my own class to do google searches. This class will do a normal
google search with the Google URI and parse the HTML to get the results.

This way I'll get the number of responses I want.