Refactoring of dc-admin-buster complete.
It seems that the run method first present itself and then check
for errors, producing a strange behavior: In case of errors, the
banner is shown but then the error message is shown and the program
dies.

This seems a unacceptable behavior for a decent tool. So, fixed it.

Future rewrites in this class are also a thought.