DoomCaster 1.9.8-BETA.
After a long time of abandonware phase, I decided to revive this project and
add a lot of interesting features to this. This time, dcdorker will have support
for multiple search engines, support for configuration of the search, the HTTP
requests made, the exploitation, etc... My objective is to create a tool that
can work as customizable, robust, flexible and decent dorker.

This is far for being complete, there's a lot of features to add and certainly
there's a lot of bugs I still didn't see.